chore(condo): DOMA-11681 implement error handling utilities and restructure error modules
